Remove code refactor lefovers
