+++ src/data.rs
@@ -6,3 +6,27 @@ pub mod elasticsearch;
 pub mod uri;
 
 pub use uri::Uri;
+
+// ------ Utility Function -------
+use crate::env;
+use color_eyre::eyre::Result;
+use serde::Serialize;
+use std::{fs::OpenOptions, io::Write, path::PathBuf};
+
+pub fn debug_save<T: Serialize>(filename: &str, content: &T) -> Result<()> {
+    if !log::log_enabled!(log::Level::Debug) {
+        return Ok(());
+    }
+    let home_file = PathBuf::from(env::get_string("HOME")?)
+        .join(env::get_string("ESDIAG_HOME")?)
+        .join(filename);
+    let mut file = OpenOptions::new()
+        .write(true)
+        .create(true)
+        .truncate(true)
+        .open(home_file)?;
+    let body = serde_json::to_string(&content)?;
+    file.write_all(body.as_bytes())?;
+    file.write_all(b"\n")?;
+    Ok(())
+}
